Add autodeployment to heroku
Added heroku requirements
Updated the circleci yaml to use the orb

Add initial logic to deploy to heroku

I think this will work.
I'm not sure if deploy will work on non-master branch.
Procfile is set to use gunicorn
runtime.txt is set to same python version in circleci

try removing comment

interesting that branches is the issue.

The circleci out says it in the passed, but then shows it
again above. Need to read the logs more clearly.

try the orb, why not

fix procfile

figured it out after running it locally a few times
or should I say failing to run it locally

Ok. now only deploy on master